requirements.txt
fastapi
uvicorn
pillow
psycopg2-binary
requests
python-dotenv
